[ADD] an alerte before the reset button

+++ script.js
@@ -3,34 +3,43 @@ var clickButton = document.getElementById('click');
 var resetButton = document.getElementById('reset-button');
 var lastscoreElement = document.getElementById('lastscore');
 var score = localStorage.getItem('score') ? parseInt(localStorage.getItem('score')) : 0;
-var flamesPerSecond = 1; // Nombre de flammes par seconde
+var flamesPerSecond = 1; 
 
-// Mise à jour de l'affichage du score
+
 function updateScore() {
     scoreCount.textContent = score;
 }
 
-// Ajout du score
+
 function addScore() {
     score++;
     updateScore();
     localStorage.setItem('score', score); 
 }
 
-// Réinitialisation du score
 function resetScore() {
-    score = 0;
-    updateScore();
-    localStorage.setItem('score', score);
+    
+    var confirmation = confirm("Are you sure you want to reset?");
+
+    
+    if (confirmation) {
+        score = 0;
+        updateScore(); 
+        localStorage.setItem('score', score);
+    } else {
+       
+        return;
+    }
 }
 
-// Fonction pour ajouter des flammes par seconde
+
+
 function addFlamesPerSecond() {
     score += flamesPerSecond;
     updateScore();
 }
 
-// Affichage du dernier score
+
 function displayLastScore() {
     lastscoreElement.textContent = "LAST SCORE : " + score;
 }
@@ -40,7 +49,7 @@ resetButton.addEventListener('click', resetScore);
 
 updateScore();
 
-// Démarre les flammes par seconde
+
 setInterval(addFlamesPerSecond, 1000);
 
 displayLastScore();
